requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
supabase>=2.0.0
transformers>=4.40.0
torch>=2.0.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
